Fix flicker after unlocking when IME was showing
This CL fixes playing the IME show animation after unlocking the screen, when the IME was shown previously.
The issue here was that we changed the control, which triggered another show behind the IME snapshot. We also refer to the unfrozen serverVisibility in ImeInsetsSourceProvider.

Test: com.android.server.wm.flicker.ime.ShowImeOnUnlockScreenTest#visibleLayersShownMoreThanOneConsecutiveEntry
Fix: 374413957
Flag: android.view.inputmethod.refactor_insets_controller
Change-Id: Iaa33ab3aeaaf1e5861a0fc890291cc1694accede

Former-commit-id: 3246c75817020fa7485de888dbd0a264b448a7b8