Allow ALLOCATE_AGGRESSIVE permission for StorageManagerService.getInternalStorageRemainingLifetime
The getInternalStorageRemainingLifetime API is used to get the remaining lifetime of the internal storage. This information can be used to make decisions about how to use the storage, especially around making many allocations. The ALLOCATE_AGGRESSIVE permission is a good fit for this API because it is a privileged system permission that relates to making many allocations in storage. A system/privileged app should only need to use the getInternalStorageRemainingLifetime API only when deciding to make a large/frequent number of allocations.

This change updates just the set of permissions for StorageManagerService. The API permission changes are in subsequent commit.

Bug: 373891491
Flag: EXEMPT small localized change
Change-Id: I31d3fed9b47024d8ffdc8c05646c7b08812dab99

Former-commit-id: b327a1806f57f7c4984fc651987b5133a84b4580