[AAPM] Add APIs for support dialog and identifiers for features
The new createSupportIntent() API returns an intent that can be used to
launch a support dialog for advanced protection features. This dialog
explains to the user that a specific action or a setting is disabled by
advanced protection.

To ensure the support dialog and the advanced protection settings screen
display the correct information, this change also adds string
identifiers for each feature and its category for the support dialog.

Bug: 352420507
Bug: 369361373
Test: atest AdvancedProtectionManagerTest
Flag: android.security.aapm_api
Change-Id: I7fa96decd2ad4403f95417c4d5e0574527b70ea7

Former-commit-id: ed62cf6727dba35cefbe3eefe2ac76d90fe37f5f