SplitScrn: Check for all children visibility when updating mVisible.
Currently, mVisible is updated based on the latest #onTaskInfoChange's
task update. However there is no guarantee that this comes "in order";
for example if the Stage task gets a separate task on top that's covering
it, it's possible the new top task comes in first, and then the old,
now not-visible task update comes next. The root is actually still
visible, but because of this order, mVisible is now false.

This CL changes it so that when we get a new task update, we check
against all children tasks' visibility to get a final value for
visibility of the root/stage.

Bug: 358065655
Test: Open Gmail and Settings side by side, and start a Compose Mail
action (generates new task) for Gmail. Drag Desktop handle - no longer
crashes.
Flag: EXEMPT bugfix

Change-Id: I44279fbc6bcda75168127e2b85c5b0c7b3918983

Former-commit-id: e984551f5c153fe7d9fc2fdda23b5cca6632e6d6