Enable smoother burn-in adjustments on transition cancel
Using transitionValue() when a transition is canceled produces
a jump in values and a bad animation. Listening to transition
edges directly does not have that same effect, as it will use
the last real value emitted. To make sure values reset, also
listent to all transition to LOCKSCREEN.

Test: atest AodBurnInViewModelTest
Test: double-tap launch camera from AOD/LOCKSCREEN
Test: repeatedly unlock/lock with sidefps
Fixes: 336241173
Fixes: 369173324
Flag: EXEMPT bugfix
Change-Id: I084f06be7133d4c9e84d5983660f5086cee76337

Former-commit-id: c9ce255f85a3c25dc3f73c82538656039b55c4e7