[SB][Screen share] Proactively start screen record timer.
SysUI screen recording has a 3 second countdown before recording
actually starts. The status bar chip shows that 3-2-1 countdown in the
chip. Once ScreenRecordingRepository says that the recording has
actually starting, that's when the status bar chip updates to show a
timer (00:00) instead of the 3-2-1 countdown.

However, there's a bit of a delay between when the recording has
actually started vs when the repository knows it's started. This means
that all the saved screen recording videos show the status bar chip with
the "1" countdown for a few frames before they start showing "00:00".
This isn't great from a user perspective, because it makes it seem like
the recording started earlier than intended (though that's not actually
the case).

This CL updates the screen recording chip to proactively start showing
the 00:00 timer based on the `millisUntilStarted` value that we get
during the 3-2-1 countdown.

Flag: com.android.systemui.status_bar_auto_start_screen_record_chip

Test: Take screen recording -> verify that the saved video never shows a
chip with "1" in it
Test: Smoke test of existing screen recording chip behavior
Test: atest ScreenRecordChipInteractorTest

Bug: 366448907
Change-Id: Iccd54153a35264a4d61a9dfe2fa482e0ec71b528

Former-commit-id: 51ef42f7a718846302ed4b74f8ab9cd3548b39af