Set a transition to ready for removeTask IFF it's the only transition.
There's a need to get removeTask calls to set transitions to be ready
right away; however this is only the case if removeTask introduces a new
transition that's only of this CLOSE change and nothing else. Since we
could rely on currently collecting transition, for that case, we will
let the transition play out normally.

Bug: 369884874
Test: Manual. Open up an activity that has activity A => B (and A closes
itself right away), and see that Desktop Mode app handle shows up
correctly instead of not showing.
Flag: EXEMPT bugfix

Change-Id: I0c53e6e6de95ea000c0ddf8f281e2308a8b0e900

Former-commit-id: d0dd29a3c75c46019200cc129520e6e257fcc2c2