Revert^2 "[SB][Notifs] Make notification chips icon-only for now."
This reverts commit ba61a33c66fb05691bd21a2a2b8eca1e743b53cb [formerly c74d1b01f914de3bfe40e044db38d30f077f6b48].

Reason for revert: Resubmitting and fixing test failure

There's a lot of work still to do on the text shown in the status bar
notification chips. Since we want to start teamfooding with it ASAP,
remove the text for now and just show the icon.

Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips
Test: Verify notification chips in status bar only ever show icon
Test: atest NotifChipsViewModelTest

Change-Id: Ie2ed7e40ca503e59ab693e3e8d78b007be6a0f8c

Former-commit-id: f2bd615ea92c274434085ea18406052ceb52f59d