DisplayInfo: don't check renderFrameRate for equal()
equal() already checks getRefreshRate in order to determine
if the DisplayInfo differs in the refresh rate the app would
experience. Checking for renderFrameRate is redundant.

This change avoids a duplicate onDisplayChanged event when
both the render frame rate and the the frame rate override change.
For example, the following are considered the same from the app perspective:
 - renderFrameRate=30; frameRateOverride=30
 - renderFrameRate=60; frameRateOverride=30

Bug: 377240935
Test: Running TouchLatency and expanding notification shade
Flag: EXEMPT bugfix
Change-Id: Ic57a785401c515f65a6e89009de43e33c949f7dc

Former-commit-id: a5fe6975828e6daa9c54684fa64a56a42a22e3d0