Extract reusable logic into smaller methods
This is a no-op change that is extracting some parts of the code in to properties and methods.
The plan is to reuse these parts of code in subsequent CLs that implement bouncer blur.
This is a separate CL to make it easier to review because these changes are not flagged.

Bug: 370555003
Test: NA
Flag: EXEMPT no-op changes
Change-Id: I7b0d6cbbe83cbbf20303731b12676ed7ee4fde00

Former-commit-id: f8dd6ef3d78594dafaa18ce3c000d110bb76d820