Implement key gesture to activate Select to Speak
Adds new shortcut (Meta + Alt + S) to first enable the select to speak
service and then activate it once it is enabled.

The select to speak service to use is controlled via the
config_defaultSelectToSpeakService value and is empty by default. This
select to speak service to use is checked to make sure it is
pre-installed from the OEM.

Bug: 375277034
Test: atest AccessibilityManagerServiceTest
Test: Manually on device
Flag: com.android.hardware.input.enable_talkback_and_magnifier_key_gestures

Change-Id: Ic22e12fd66010890d6d495b6c52a52008dc0ca9e

Former-commit-id: 1a239822e1461d856fe2f4fd844f09ed65def40e