Fix AE LockTaskMode issue on fold devices
When set to Lock Task Mode on outer screen, the embedded activity is always displayed in full screen. This results in this behavior: left activity still full screen, placeholder is launched on the right.

It seems that "dropping all multi-window windowing mode change" is not as expected in AE for fold devices. Skip this judgment for AE can solve this issue.

Bug: 376366801
Change-Id: I4c9963883f8a4886eda5cccaac646bf77ab12db8
Signed-off-by: biance <biance@xiaomi.com>

Former-commit-id: e5c40386e5e164020f9a06ed7ec3fb16a98c5944