Add client-side APIs for getRoutedDevices()
After you open an audio stream, you can call getRoutedDevice() to
get which output device is actually used.

However, if you play an ALARM with a headset attached, audio comes
out of both the speaker and the headset. This is not properly reflected
with our current APIs.

This CL adds the client side APIs for getRoutedDevices().

Bug: 367816690
Test: presubmit
Flag: android.media.audio.routed_device_ids
Change-Id: I23ba7c4a3dcaa9683db638778bdff65fe422f8df

Former-commit-id: a1be72756c4dbb7eb5c1e480a869ccb1518d9139