[SB][Notifs] Hide status bar notif chip if the app is open.
If the app that posted the promoted notification is visible, we don't
want the status bar chip to also show in case the information in the
chip is out-of-sync with the information in the app (e.g. a timer is
slightly off).

OngoingCallController already does this using a UidObserver that it
manually registers and unregisters. This CL does things a little
differently:
1) Use a new ActivityManager API instead of an IActivityManagerAPI
2) Create a repository responsible for listening for UID updates
3) Use `conflatedCallbackFlow` inside the repository which should handle
   the registering and unregistering automatically
4) Have SingleNotificationChipInteractor use that repository to create
   the flow

Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips
Test: Start a promoted ongoing notification, then open the app
associated with the notification -> verify status bar chip disappears.
Close the app -> verify status bar chip reappears.
Test: atest SingleNotificationChipInteractorTest

Change-Id: I3012b6a99e723e7e0f0f921fbb8bde2aabb5857e

Former-commit-id: acc477c4eff08896d58ca9aaa3802e19a31bb094