Make the transition DSL more opaque
This CL makes most of the classes in the transition DSL internal instead
of public.

In the past, the TransitionSpec's returned by the from() {} and to() {}
builders were used, but it's not the case anymore. By removing these
returned specs we can make most of the DSL classes internal.

Bug: 376438969
Test: atest PlatformComposeSceneTransitionLayoutTests
Flag: com.android.systemui.scene_container
Change-Id: I1b4e37b4c105d5f2253629ddd96fdfc0e96361fc

Former-commit-id: 93747b75d73043fcf31200d4b774576f871aac80