Do not invalidate isInteractive cache on power group events.
ag/29700291 fixed a bug with the cache invalidation as it didn't take
into account individual power group wakefulness changes. However, it
also invalidated the cache for any power group event, which was
unnecessary - when a power group is added, changed or removed, the
cache should be invalidated only if the global wakefulness has
changed. So reverting that part of the CL.

Fix: 372215504
Test: presubmit
Flag: EXEMPT bugfix
Change-Id: I52a135a02008ac50c403715729960b9887a581c7

Former-commit-id: 2fea7634cc0c4c36e2ba118e239168ef1db67c0f