Ignore ASHA hisyncId grouping if the device supports CSIP
If a device is added as another device's sub device, it'll be removed
from the cached device list. This makes the device can't be correctly
grouped as a CSIP set when CSIP is connected.

We should ignore ASHA hisyncId grouping and leave it to be handled as a
member of CSIP set.

Flag: EXEMPT bugfix
Bug: 376022366
Test: atest HearingAidDeviceManagerTest
Test: manually check the issue is resolved with real device
Change-Id: I5897bcbd1915d03157743ba09e4092c6976736c3

Former-commit-id: 214032f16b529fc2cba735705d459156cb1ca9c8