[flexiglass] Prevents crash when disable flags are set
There's a crash when certain disable flags are set. It comes from the
measure policy of SingleShadeMeasurePolicy which assumes that certain
elements are laid out when they're not (because the disable flags makes
those elements disappear in runtime).

This CL only prevents the crash, it doesn't fix the positioning of the
UMO in these cases, that still needs to be tested and fixed properly.
See the test lines for how to fix it properly. This is tracked in
b/379162317.

Bug: 371545613
Test: manually verified that disabling certain disable flags doesn't
crash anymore. The commands used are:
$ adb shell cmd statusbar send-disable-flag quick-settings
$ adb shell cmd statusbar send-disable-flag statusbar-expansion
$ adb shell cmd statusbar send-disable-flag none
Flag: com.android.systemui.scene_container

Change-Id: Ie26bbb743f06e2eb1549e67b9a5a3c7ca0613857

Former-commit-id: 2ccd4af147b6ad30b5e855fc86083d9b916e52e3