Use addMoveToFullscreenChanges on fullscreen launch.
When a desktop is minimized and a task contained inside it is then
launched in fullscreen, the windowing mode is always set to
WINDOWING_MODE_FULLSCREEN. However, when the display windowing is
fullscreen, the task should be set to WINDOWING_MODE_UNDEFINED to
prevent issues with future split screen transitions (see
ag/23624076 for more details).

This is already done for other freeform -> fullscreen transitions, this
CL just applies the logic to task launches.

Bug: 372809656
Test: Manual
Flag: EXEMPT bugfix
Change-Id: Ib05f7b9912d3cc3b834b9e3217a56337de3c163d

Former-commit-id: 6b4d6ed074fd7c19a6f9074e36a84db26fefe28f