Handling intent whose nested intent keys not collected case
We have tried to cover all the public surface of APIs to always
collect nested intent keys. But there could still be cases where
prvileged app calls hide API's that send an intent over IPC but
does not call collectExtraIntentKeys() before. This will cause
those nested intents failed to be launched because they will be
marked as missing creator token. In this CL, we will either throw
an exception when this happens, or collect the keys on the server
side, depends on which flag is enabled.

Bug: 379366027
Test: Manual test.
Flag: android.security.prevent_intent_redirect_throw_exception_if_nested_keys_not_collected
      android.security.prevent_intent_redirect_collect_nested_keys_on_server_if_not_collected

Change-Id: Idb3136b22a892252def272b2f48e085d38aec5fe

Former-commit-id: b542fa0104c4c0e7f4b816d90364501d36080a6f