[AAPM] Add DevicePolicyManager#getEnforcingAdmin for settings pages
With advanced protection setting various policies, the relevant device
policy strings and support dialogs need to be updated to state a certain
action was blocked by advanced protection.

Currently, the device policy dialog determines the enforcing admin by
calling the existing getEnforcingAdminAndUserDetails method. This method
returns a Bundle, which is not enough for the dialog to infer that a
policy was set by advanced protection. Hence, this change introduces a
getEnforcingAdmin method that returns an EnforcingAdmin object with a
correct Authority.

This change also updates existing RestrictedPreference files to use the
method and display correct strings in the switch summary.

Bug: 369361373
Test: manual
Flag: android.security.aapm_api
Change-Id: If931dcddad508f88aac1280b587da4767b937875

Former-commit-id: e9727237d81745d5b61044a569ff81f0c5ba0604