Remove special case for replaced transitions
This CL removes a special case that was added to slim down the
STLState.currentTransitions when replacing a transition by another one.
This optimization was unnecessary and adds a mutation of
transitionStates where removed transitions were not logged as finished.
This premature optimization is not worth its cost given that it had
already caused a crash in the past that was fixed in ag/29881406.

Bug: 379281707
Test: atest PlatformComposeSceneTransitionLayoutTests
Flag: com.android.systemui.scene_container
Change-Id: Ide9fa14a9f1883b5a13ee792d84ef3a58a6923cc

Former-commit-id: 8bd1ed8052f329f67bb9a6974d5341664bebaa09