Add pip2 flag check into some WM tests
Some of the test cases are only relevant to PiP1.
For example setting activity windowing mode independently
was only being used as a hack in PiP1, so require pip2 flag
to be off for this test.

We have also been having issues with flag override in this test
after teamfood rollout, so use RequireFlagDisabled instead of
DisableFlags, similar to testLockAndUnlock() (see b/379365576).

Bug: 379365576
Bug: 379365059
Flag: com.android.wm.shell.enable_pip2
Test: atest WindowOrganizerTests
Change-Id: I7da761e1203f61f49e0854a8679f9c60e6998430

Former-commit-id: f0528c3c4f31d1d7450d6e37966b394c884a52fd