Ensure we bind to the correct app's agent
We have a 10s timeout in bindToAgentSynchronous and if we don't get an
agentConnected callback by then the method returns null. There was a
potential bug here because agentConnected doesn't check if the agent is
the one we are currently expecting. It could happen with the following
sequence of events:

1. bindToAgentSynchronous(A)
2. app A times out
3. bindToAgentSynchronous(B)
4. agentConnected(A)
5. the call from 3 returns the agent for app A instead of B.

With this change we start tracking the state of the current connection
better.

Test: atest BackupAgentConnectionManagerTest & atest
CtsBackupHostTestCases
Bug: 328019933
Change-Id: I0556aff00fe8c2bbfe56fe10f15b7b15ff260590

Change-Id: I4c04e9279d9863716fc080c6ab218665e1591d51

Former-commit-id: 64d73c4f0360c35e8a4b8180778ecf65aef5c2cc