Remove invokeOnImeRequestedChangedListener call if input target != control target
In case when the input target was different to the control target, we already invoked IME listener (in ImeInsetsSourceProvider#reportImeInputTargetStateToControlTarget). However, this should only happen, when the caller of updateClientVisibility is the control target.
In split screen, this will happen (through a call to WindowManager#updateDisplayWindowRequestedVisibleTypes). Therefore, we don't need to trigger the listener too early.
The virtual display is an exemption, as is not guaranteed that it has an ImeInsetsSourceProvider.

Test: CtsInputMethodTestCases
Bug: 298172246
Flag: android.view.inputmethod.refactor_insets_controller
Change-Id: I6fe7e2e1f972be264796e8880c006f15aa18f408

Former-commit-id: 13d5dce50e1753d6aac44bcaf1b21e576022d35e