[PiP2] [2/2] Start with a non-null pip params
We need to make sure PiP params are not null when
swiping home to PiP as well. This is because a client can
call into enterPictureInPictureMode() with an empty PiP params
which Core's fillTaskInfo() will resolve as a null pip params.

So always operate on a non-null PictureInPictureParams to avoid NPE.

Do the same for expand animation too, when querying cached params.

Bug: 376141728
Flag: com.android.wm.shell.enable_pip2
Test: swipe-pip-to-home with an invalid src-rect-hint
Change-Id: I7bb1158bd0593788a1e1514f538c0a433841ad61

Former-commit-id: 0508c7f409c37e6b125f2416d0286499f7ba2663