Require PiP2 flag off for dismiss split test
Overall the tests are flaky and have been failing
before the enable_pip2 flag rollout to staging.
But it looks like the [ROTATION_90_GESTURAL_NAV] config in particular enters PiP
when exiting splitscreen with the primary app,
which means this is hitting the DefaultMixedHandler + PiP2 path
which we have been planning on refining after trunkfood-partial.

So require enable_pip2 flag to be disabled to run
DismissSplitScreenByGoHome flicker test

Bug: 380036543
Flag: com.android.wm.shell.enable_pip2
Test: atest DismissSplitScreenByGoHome
Change-Id: I6c60ee0c8bd607989aff31fed4b12aca0c76aa08

Former-commit-id: 8caa1ad408720aacf293df8bfd0689c7e1f7af1f