Remove WRITE_DEVICE_CONFIG permission from shell user
Android 16 adds restrictions to the DeviceConfig flags that can
be modified by adb shell; to do this, the WRITE_DEVICE_CONFIG
permission will be removed from the shell user, and it will
instead rely on the WRITE_ALLOWLISTED_DEVICE_CONFIG permission
along with the allowlisted namespaces and flags. This commit
removes the WRITE_DEVICE_CONFIG permission from the shel user's
manifest to ensure it is not able to modify all DeviceConfig flags.

Bug: 364083026
Flag: android.security.protect_device_config_flags
Test: atest DeviceConfigApiTests
Change-Id: I7878d9988ca1223be2c897c0d491c1cdb2320395

Former-commit-id: d14211317345447f37f0061c92ce58fd40b152aa