Adding conditional fallback to V implementation
Older tests that rely on AppCompat flags may still be executed on newer code versions. We need to retain support for the older behavior when the feature flag is disabled. This change reintroduces the previous implementation method without modifications. This method will be called if the feature flag is not enabled, ensuring that older tests continue to function correctly.

Bug: 378656966
Test: atest FrameworksCorePackageManagerTests:ParsedMainComponentUtilsTest CtsPackageManagerTestCases:SaferIntentTest PackageManagerServiceUnitTests
Flag: android.security.enable_intent_matching_flags

Change-Id: Iaad7fac084c317b128b3beefe72914e03aceac30

Former-commit-id: 07618c481d6611e634b898165dd27f21a61d469b