Introduce ShadeDisplayPolicy
A policy defines how to select the display that should show the shade.

For now, the policy is set only by the the adb command.

In this cl only the "static display id" policy is introduced. See child cls for more policies.

Bug: 362719719
Bug: 380444270
Test: ShadePrimaryDisplayCommandTest, ShadeDisplaysRepositoryTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I01abbef21f3c386aad169768c9cd01533dc4ecd2

Former-commit-id: 3fa30d350b56bbaf6f916b902397afbc4c67a0e3