Update communal db for responsive grid
In the responsive grid, we are updating the grid cells such that a
widget occupies a single cell instead of the 3 cells they do today. We
therefore add a new spanY field in the database which stores the new
height in responsive grid units.

Also noticed that we were not properly backing up the spanY field
before, so fixed that in this change also.

Bug: 378171351
Flag: com.android.systemui.communal_responsive_grid
Test: atest SystemUITests
Change-Id: I147e1ba5f7fa1c0ad2eaffcd02b11bfd062ab55d

Former-commit-id: 81b1558afb9e144a8b369cfe13d755d8ea74f6df