PiP2 DW: Do not enter pip if opening a new task while in desktop mode.
In 3-button navigation, when we enter PiP we expect a TRANSIT_OPEN
transition request. However in Desktop Mode, since we always open in
fullscreen first and then shortly transition to Freeform, there will be
a short window where Core will treat this app launch as a normal
fullscreen launch and do all the normal lifecycle things, including
trying to Pause -> enterPiP. This is also a TRANSIT_OPEN request, so we
end up PiPing even though we do not want to. Putting an explicit check
here to stop PiP.

Bug: 351133000
Test: Manual
Flag: EXEMPT bugfix
Change-Id: If34483f4a2eaf2af5f9c9a05bed5867455187799

Former-commit-id: 1b74733ad6957c762bdf0fba66a576527b32a027