Revert^2 "A few fixes for animation takeovers."
Reason for revert: this test on Cuttlefish seems to be very unstable in general. I tried to re-test on ABTD with these reverts patched in, but the test isn't being included in the suite, and I couldn't find information anywhere on whether it's supposed to be still running. Since these changes are fully flagged, I don't see how they could be the real culprit. Trying to submit again and monitoring the impact on the post submits.

Change-Id: I909b4e2c00296ca7ea66f73ad25f843ae64aad40

Former-commit-id: b6557ed79bedf5ae242032eb9d0654f8a2967b0d