Use new tiles in TileRequestDialog
This adds a flag check to TileServiceRequestController to check for QS
in compose. If true, it instantiates a compose dialog (with same content
as before). We also make sure that the icon is loaded in the background
thread.

Test: manual, using test app
Test: atest com.android.systemui.qs.external
Fixes: 369576789
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Flag: com.android.systemui.dual_shade
Change-Id: Idc4f30c503b08bdcba461931da291f59501bc12c

Former-commit-id: 9eb44a88ed909899e72c048ed0b52d3ead8cfd3c