Demote top-app when notification shade is expanded
If the top app is busy, it may cause potential jank when expanding
notification shade because CPU contention.

If there is top app switch, the demote will be canceled. Because it
is usually an app launch which needs top priority for startup latency.

Bug: 362467878
Flag: com.android.window.flags.scheduling_for_notification_shade
Test: Expand/collapse notification shade and check
      adb shell dumpsys activity o | grep -e "A.*top-activity"
      Expanded: intermediate-top-activity
      Collapsed: top-activity

Change-Id: Ic82cb3235f93f01db382d355c98f7aef2f1ef783



Former-commit-id: 4f26be3a00ba220b427832e229b79ecbc57b57c1