Adding information about whether a quick affordance launches.
Previously, it was difficult to disambiguate whether a quick affordance
handled an action by simpling toggling a control or by launching an
activity or a dialog. This information is crutial for the delivery of
semantic haptics via the Multi-sensory Design Language (MSDL). To supply
this necessary input, the Handled result in
KeyguardQuickAffordanceConfig.OnTriggered result now includes a boolean
property to determine if the action was handled by launching.

With this, the new data class LaunchingFromTriggeredResult consolidates
the launching result of an action on a quick affordance, regardless of a
launch occurring via the system or the quick affordance itself. The
subsequent CL uses this information to deliver the correct
MSDL-compliant haptics.

Test: KeyguardQuickAffordanceInteractorTest
Flag: com.android.systemui.msdl_feedback
Bug: 361320572

Change-Id: Ifc42748589cab885059a24a83537e5496e4fbf40

Former-commit-id: 7415da587004876bef28332721e9fb0a2c5d8ab1