Fix incorrect flag for starting DynamicInstrumentationManagerService
SystemServer was previously referencing the flag
com.android.art.flags.executable_method_file_offsets, which is in the
ART (mainline) container, and can thus be set to true on pre-25Q2
devices.

The relevant SELinux policy for DynamicInstrumentationManagerService is
only present from 25Q2 onwards. Thus, the system would crash on pre-25Q2
devices with the ART flag turned on.

Correct this by useing the flag of the same name in from the
android.uprobestats package, which is in the system container and will
only be enabled from 25Q2 onwards.

Bug: 382532566
Test: TH
Flag: android.uprobestats.flags.executable_method_file_offsets
Change-Id: I66bae90e49609cd3ee7f7e7b9a64e8f82e4b8641

Former-commit-id: bde7cf67464b8b1b8d47b930265f4aa4cbd1fa95