[SB][Notif] Use PromotedNotificationContentModel.time for chip time.
`PromotedNotificationContentModel` has a `time` field, so we should
re-use that field instead of having a custom `whenTime` field.

This CL also updates the chip to not show the time if
`Notification.showWhen` is false (if `showWhen` is false, then
`PNCModel.time` will be null).

Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips
Test: Post RON with showWhen=false -> verify chip just shows icon
Test: Post RON with showWhen=true -> verify chip shows time difference
("4min" for 4 minutes in the future, "1hr ago" for 1 hour in the past)
Test: atest NotifChipsViewModelTest SingleNotificationChipInteractorTest

Change-Id: I43d2c6cf2bd70ef2232cbd66357893e201a4f8dd

Former-commit-id: cdbccfa3ff9113b14828d3686b6e77a6f7b61340