Refactor native store for PIC nonces
This change shuffles code for the native PIC nonces.  Implementation
details are moved out of the native PIC header file.  This is in
preparation of adding futex operations on nonces; we do not want all
the logic for futexs in this header file.

The native class is also broken into a base class with methods and a
template that creates a subclass with variable storage.  All methods
are against the base class, which means that modules like
ApplicationSharedMemory and PropertyInvalidatedCache can talk to each
other in terms of the base class.  One day, we may have more than one
instance of the nonce block.

Flag: EXEMPT refactor
Bug: 373752556
Test: atest
 * FrameworksCoreTests:PropertyInvalidatedCacheTests
 * FrameworksCoreTests:IpcDataCacheTest
 * CtsOsTestCases:IpcDataCacheTest
 * ServiceBluetoothTests
Change-Id: I7155898581ea98178ab80c07b4b2d5a8f2df2a16

Former-commit-id: 54cc588f32d62b5916928374ea432d71f1e9e9cd