Fix missing creator token issue due to same intent added twice as extra intent
If the same object is added more than one time to a top level intent
as extra intent, but with different keys, collectExtraIntentKeys()
method would only collect one of the keys because it considers the
second intent as a duplicate and not handling it again. If later,
the app retrieves the intent through the second key and tries to
launch it, this would result as a failure due to missing token error.

To fix it, we should still collect the key even if the intent is a
duplicate; we just have to not handle the intent recursively when
we detect it as a duplicate.

Bug: 382863447
Test: manual
Flag: EXEMPT bug fix
Change-Id: Id3cbf15ef1902167ef52068b27765b4aa942ab8b

Former-commit-id: 20b229794ace3f58d5ec791a57e48362da319dd1