Remove flag check enableOnDeviceIntelligenceModule
This flag check fails for 25q1 release build as there the module migration build flag is not ramped, and this method is accessed from services.jar itself. There is no need to check for this flag for 25q1 atleast, for 25q2 since we are ensuring manually, the API flag is ramped before the build flag, that works fine too.

Change-Id: I93c75270e3130fea1cce676aa5338fcaea7ca3b2
Bug: 382429792
Flag: EXEMPT bugfix

Former-commit-id: f2a006ffd7810914fb25e974ac463c9583659b7a