Provide @Main decorated context for the default display
This creates a binding for the default context using the @Main annotation.

As we're now using several different types of context in sysui (e.g. a shade window specific one), classes that are aware they need the context only for the default display should be annotating it directly with @Main.

This also ensures consistency as we're using @Main for several other classes.

No injections of undecorated context will be forbidden with a linter eventually (and is already forbidden in shade related packages as of today with ShadeDisplayAwareContextChecker)

Bug: 362719719
Bug: 374267505
Test: presubmits
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I3b0c1bc1659a2285c5ca55dd10ed8d2d2d0b7abe

Former-commit-id: 79079b80076d40e545e61991e12958072f70e60d