Fix Order of Logging
Our metrics record duplicate and default values. This can be solved by
moving the order of the emits further down. The method also needs to be
locked so that it is not called multiple times - or if so, is not facing
any race conditions that may continue to cause unexpected outcomes.

Bug: 362995084
Bug: 362994633
Test: North stars in metrics counts - repro for this locally has been
attempted, though only collected metrics have this issue
Flag: android.credentials.flags.fix_metric_duplication_emits

Change-Id: Iae0d31d3c5b48897fe72b69bb26b09c3d35a756b

Former-commit-id: 52d6bcb8d32ef83f7b22eb64dc53eec11e606982