Fix flash of QQS when collapsing in Keyguard
The issue was that qsVisible became false (same for
clippingParams.isEnabled), so QQS was not being clipped anymore, but the
AnimatedVisibility would then play its exit animation making QQS visible
for a frame. This was not observed outside of lockscreen because there,
QQS is always marked as visible (this is an upstream issue that we
resolve by also checking if the shade is expanded).

To solve this flashing issue, replace the AnimatedVisibility with an if
block. This is not an issue because all animations for QQS/QS (like
alpha) are controlled externally.

Test: manual, fling collapse QQS in lockscreen
Fixes: 380416826
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Change-Id: I2db2a4f365ad74db9971cc0d11638918a3fbddd6

Former-commit-id: 7f65a2e216e9511d0f9a917a6671cb449a8e65db