Add builder pattern to build WindowState in WindowTestsBase
The current `createWindow()` method has > 10 overloading
patterns, and many calls with null as params as they're not
needed, and as to not add more overloading.

In the same WindowTestsBase, there's also been TaskBuilder,
ActivityBuilder, TaskFragmentBuilder, and therefore having
"WindowStateBuilder" also fits the class here.

Refactoring of all the usages will be done gradually for
each individual test classes.

Note: No test would be impacted by this as this is adding
a new helper class

Bug: 383480021
Test: presubmit
Flag: EXEMPT cleaning up tests
Change-Id: I56296d595969325a0d3cd63252b8ec5db7e56ed9

Former-commit-id: ab1182c33a05b243384a3d8b86b8d1569ec6d1ec