[flexiglass] Moves NSSL up on wide lockscreen scene.
On unfolded foldables and tablets, the NSSL was positioned too far down.
The reason was twofold:
1. The padding was duplicative, the NSSL is already being placed in the
   right spot in the layout, which makes it take account for the status
   bar height (it's placed below it on the Y axis)
2. The padding value was incorrect. A px amount was treated as dp which
   incorrectly showed a multiple value of the height, exaggerating the
   bug

The fix was simply to set the padding to a minimal 12dp which makes it
match the position of the NSSL on pre-Flexiglass layouts.

Fix: 377903994
Test: manually verified that the NSSL position is the same with
Flexiglass on and off on both an unfolded foldable and a tablet
Flag: com.android.systemui.scene_container

Change-Id: I9d06cf2df984b9e40dd64b851f638ba2497c47bf

Former-commit-id: f40858f2ed23d1337ff30d2ddf578796a8aafa56