Propagate onMovedToDisplay with ConfigurationRepository
This propagates onMovedToDisplay signal from ConfigurationController to ConfiguratoinRepository, providing a flow that emits the latest display.

For now (and probably forever) the new configuration is neglected, as it's already available from the configuration flow. If it turns out useful in the future, it will be added

Bug: 362719719
Bug: 378688537
Bug: 378686382
Test: ConfigurationRepositoryImplTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Ie369b7e89a158cae15b463ccd2806ec4869ae08d

Former-commit-id: 5f4a8c13f046e737df1b1bd2f3131f4c8e27692c