Introduce ShadeTraceLogger
This class unifies trace events related to the trace in a new track in perfetto.

This makes understanding what's going on much easier, as events are not lost in the middle of the main thread, but immediately visible and easy to parse.

This will be particurlarly helpful to debug cases of slow shade move between display from field traces.

Bug: 362719719
Bug: 378688537
Test: perfetto trace
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I16f037945388e810fa21b5436cffab26e87a8f40

Former-commit-id: 87f0e837a46253f61af07c3baf03fee90311fbb8