Restrict the symbols exported by libandroid_runtime on host
This explicitly defines which symbols should be exported by
libandroid_runtime on host. This is to avoid potential collisions with
other libraries loaded on a host system.
This ensures that the symbols needed by Ravenwood are still exported.
This also fixes the dist path of the builds.

Flag: NONE host-only change
Bug: 382086021
Test: m libandroid_runtime on host
Change-Id: I51c9b3520f3b271ab413d233c772708a1ac1af09

Former-commit-id: 14f082e42209b630c81957ae074879797ff7b239