Add hidden broadcast to inform system apps about changes in user info.
Due to new feature of freezing apps including system widgets, there is
no updates on system ui when user change profile picture or name.
Because some profile pictures are synced in in background we need to
make sure we update system ui when it happens.

This broadcast is protected by special permission and can be only used
by preinstalled system apps. The broadcast will not slow down the system
as it is fired on rare occasions.

Bug: 383071252
Flag: EXEMPT bugfix
Test: manual
Change-Id: I97c63d27040c6a313164e657d9bbb62f66cb4a3f

Former-commit-id: 55319db9308080f151d5377aa5d54e65e88a98bb