[cli] allow for an empty arg list if nothing is required
Before, we treated an empty arg list as an error. I don't think this is
necessary though. We can only invalidate an empty arg list if there are
non-optional parsers attached.

This allows for a completely optional command / subcommand structure
that still calls `execute()` when given the empty arg list.

Test: ParseableCommandTest
Fixes: 384947066
Flag: EXEMPT bugfix
Change-Id: I908c63bae089373f7b2181c0caae23eaa5f6022b

Former-commit-id: 1cf2557e7ed7beeea867f5dbc3e660f89280292a