Reset IME token after updateSystemUi during unbind
When running tests we initialze and start a test IME, which is unbound
at the end of the test. Unbinding the IME should call
updateSystemUiLocked to update the back button disposition mode, and the
IME Switcher button visibility, but this call early returns, as the
current IME token is reset just before, leaving the navigation bar with
the stale state.

In [1] we accidentally started reseting the current IME token before
calling updateSystemUiLocked, which made it essentially a no-op. This
fixes the issue by restoring the correct ordering.

  [1]: Ic3aa02925fe460d4530e9f0cede3aa3a98a109ec

Flag: EXEMPT bufgix
Test: change to three button navigation mode
  run `atest SpellCheckerTest.misspelled_easyCorrect`
  observe no IME Switcher button in navigation bar after the test
Bug: 384951386
Change-Id: I952b45b24d8f46db099fc2c72a3ec7a47258295b

Former-commit-id: ebfccdb2a36cdc7af49e5a864babbd8769ee5a5e