Remove unnecessary build steps for ravenwood-runtime.
- Use framework-platform-compat-config created from the host jar,
which is faster to rebuild because it won't have to run D8.

The generated XML is identical. See the inline comment for more details.

We should be able to use it for the device side build too, but we
don't do so 1) to avoid the risk of subtly breaking something
and 2) it's unclear if the D8 is actually a bottle neck for non-ravenwood
workflows too.

- Do the same thing on services-platform-compat-config, but
also use the version created from services.core.jar, instead of services.jar.

This is to avoid dependency on services.permission.jar, which is very
slow to build.

Currently Ravenwood only uses services.core.jar, so the compat-ids from
the other services.*.jar modules are not needed.

Flag: EXEMPT host test change only
Bug: 384791418
Bug: 384938854
Test: $ANDROID_BUILD_TOP/frameworks/base/ravenwood/scripts/run-ravenwood-tests.sh -s

To make sure the generated XML is identical...
Test:   cd $ANDROID_BUILD_TOP/out/soong/.intermediates/frameworks/base/ && diff framework-platform-compat-config/android_common/framework-platform-compat-config_meta.xml  framework-platform-compat-config-from-host-jar/andr..

Test: Manual test -- make sure out/soong/compat_config/merged_compat_config.xml
    doesn't change with and without this change.

Change-Id: Ie28eefe9726cb9011c6518d251acafd29cf1f995

Former-commit-id: 2ddb31a27c518f740354e1e9c63e5a6073e71739