Wait for overscroll to finish before finishing a transition
This CL ensures that a SwipeTransition waits for the associated
overscroll effect to finish its animation before the transition is
flagged as finished to the STLState (and therefore removed). This avoids
recomposing the STL right when the user release their finger, thus
avoiding potential jank in the middle of a transition.

Bug: 383524412
Test: Manual, saw that the recomposition happened at the end of the
 transition
Test: Performance runs in Forrest, see http://screen/3KjNsghkEUm9nVs
Flag: EXEMPT performance fix
Change-Id: I5ed31ed94a99210c144497b630265c41548f662c

Former-commit-id: b8cd7a7627f7c0aa4113b45c7772c0e07d8bb8f5