Port WindowChangeAnimationSpec to shell transitions
This was the default animation for window size changes
in legacy. This change ports it to Shell so it can
be used by the default transition handler for the same
purpose.

Also added support for animating (surface)view in
addition to the default surface-only version.

Bug: 384976265
Test: Animation-change, so existing tests.
Flag: com.android.window.flags.port_window_size_animation
Change-Id: Ibee129aa26371d32b14cb1ded9973b55929634d7

Former-commit-id: 4365fb257dcfc779ed9efdb8265d6f8d3b6ad382