[flexiglass] Fix the scroll offset in Single Shade
When nested scrolling is enabled, the scrim is always offseted at least
by the collapsed header height. Add this as bottom padding to the scrim,
so it will match the visible scrim height when its fully expanded,
and we will able to scroll through the entire stack content.

Bug: 380210512
Test: check the bottom padding of the stack in single shade
	- empty shade (no notifications)
	- few notifications (not enough to have a full height shade)
	- lots of notifications (enough to scroll the shade)
Flag: com.android.systemui.scene_container
Change-Id: I125a22becd29c70f19fa8d954e8c4076fa3920b3

Former-commit-id: caa6c158753a2064bcb607920567fb9232dd4a5a