Make the metadata section clickable only if the intent is available.
If there is no application package name provided to the dialog, or if no intent can be generated, don't add a click listener to the metadata section.
Adding a click listener sets `clickable=true` property to the element and makes it discoverable by accessibility instruments like Talkback or Switch Access.

Bug: 377167322
Flag: EXEMPT bugfix
Test: atest MediaOutputAdapterTest
Change-Id: I62c5dd6e465a76e7ae748250ff8c4a564c8d6e55

Former-commit-id: f2662cc18118df4a0ceaeafcc50156ca19af4cdd