[ROSP] Add a container for StatusBarPopupChips to the status bar
This CL adds a composable container in status_bar.xml that will
be used to display StatusBarPopupChips. The container currently
shows the song name as plain text.

Also, adds support for the MediaControlChip in
StatusBarPopupChipsViewModel. The
`StatusBarPopupChipsViewModel.shownPopupChips` StateFlow is
used in StatusBarRoot to update the `StatusBarPopupChipsContainer`.

Test: Play media -> Verify the song name is displayed on the top right
in the status bar.
Bug: b/385245633
Flag: com.android.systemui.status_bar_popup_chips
Change-Id: I5bc4e28104bacbf0846a7e5527fa399ae64607ad
Former-commit-id: 04ea67051c9ac6fa599f78be5b60ceba3c855397