[Notif redesign] Center top line vertically when needed
The top line text and expand button are usually aligned near the top of
the app icon, but in some cases where we're only showing this line of
text in the notification with nothing below it, it should be vertically
centered instead.

More specifically:
- in group headers when expanded
- in automatic public views
- in minimized views, including low priority groups

Also added some explicit nullability annotations in
NotificationChildrenContainer because I originally hadn't realised that
the group header can be null and got an NPR. This should help the IDE
warn against that in the future.

Bug: 378660052
Test: manually tested all the cases specified above, screenshot tests to come later
Flag: android.app.notifications_redesign_templates
Change-Id: Ic08ec354a435123dd8e58bf25fc58e78d246d47c

Former-commit-id: 4da4efc9aa423dbc1c480501f6f28d7d61e60bb5