Refactor WCT for null and type safety.
This commit refactors the WindowContainerTransaction (WCT) class to
enhance its maintainability through improved null and type safety.

Specifically, this includes the addition of @NonNull and @Nullable
annotations to method parameters and return values throughout the class.
The @ChangeMask and @HierarchyOpType annotations are introduced to
enforce type constraints on fields that represent bitmasks and enum-like
integer values, respectively. Finally, constants are used instead of
magic numbers to clearly conveys the intent of certain checks.

Bug: 387193964
Test: atest WmTests:WindowContainerTransactionTests
Flag: EXEMPT refactor
Change-Id: I482b46641ba9e2cce840c2934cfe066872de4db4

Former-commit-id: 93ed5aa58678d681acc5b8a8c7e32996ea2ebd42