Stretch and rescale segments to satisfy minimum width requirement.
The same function also returns the location of the progress on the
rescaled progress bar, which is used to calculate the tracker position.

The fallback option (for when it's impossible to satisfy the segment
minimum width) will be implemented in a separate CL following this one.

Flag: android.app.api_rich_ongoing
Bug: 372908709
Bug: 383566179
Bug: 383564913
Test: NotificationProgressBarTest
Test: Post ProgressStyle notifs via test app
Test: patch ag/30006048 and run ProgressStyleNotificationScreenshotTest
Change-Id: Ieb551182a9823d7231db345fcdf4a8d09ad4acaa

Former-commit-id: ef7cce266ca0592f7f7cef61c63e2eea56552393