Monitor start transaction commit time of transition
The commit time approximates the latency from the invocation of
ITransitionPlayer#onTransitionReady by core until shell applies
the start transaction. Which usually stands for the cost of
default TransitionObserver, TransitionHandler that are called in
Transitions#onTransitionReady. In other words, it is similar to
the delay between the app drawn and the start of animation.

The source is from I9845f17bcf2e0abead7299f1900d219d3ff2f66f

Bug: 370952372
Bug: 206872204
Flag: EXEMPT test
Test: atest InTaskTransitionTest
Change-Id: I2fcd1ba85137489e8118c75c6d16205ad46f5fbd

Former-commit-id: 6045a12dfeec728b028f6e135550b1da04864c6f