Use hasSize instead of size.isEqual
This will print out the list on failure, offering more information
for diagnostic.

The tiles_stableWhileCollecting  test fails less than 1% of the time in
CI with an unexpected number of tiles. Since I can't reproduce the flake,
let's at least add a bit more information to the error message for the
next time it happens.

Bug: 387159145
Test: ran test with a different expected value to check the error
message
Flag: TEST_ONLY

Change-Id: I8e4e74eeadea02d7bcd88b738cb6f3c6aff60333

Former-commit-id: 9bebaa5c18373453fb3718b6cdda16cc845fb5cf