QSDetailedView: Only dismiss screen record dialog in the dialog
The functionality to dismiss the dialog should be in the DialogDelegate,
rather than in the view binder, since the detailed view will not dismiss
the dialog when start recording button is clicked. Thus, this CL extract
the part to dismiss dialog out and move it back to the DialogDelegate.

This CL follows the refactor mentioned in
go/al-screen-record-detailed-view.

Bug: b/378514312
Flag: NONE refactor
Test: ScreenRecordPermissionDialogDelegateTest,
SystemCastPermissionDialogDelegateTest,
ShareToAppPermissionDialogDelegateTest
Test: Click on screen record tile in the QS -> verify that screen
recording in single app and entire screen both works

Change-Id: I028f1d3c1a3fb62363e432504a51eda72ca29aa7

Former-commit-id: baa89426232956b704783cff13ccbb5c7bbb8050