Include bitmap memory estimate in AppWidgetServiceImpl dumpsys output
This change adds a new field "view_bitmap_memory" to the
AppWidgetServiceImpl dumpsys output, which holds the estimate of bitmap
memory used by a widget held in memory by the service.

To support this, a new method RemoteViews.estimateTotalBitmapMemoryUsage
is added, which includes memory from both bitmap and Icon actions.

Also updates RemoteViews.estimateMemoryUsage to return a long. The
int sum can be overflowed to beat the memory check in
AppWidgetServiceImpl.

Bug: 377356756
Test: RemoteViewsTest#estimateMemoryUsage_*
Flag: EXEMPT bugfix

Change-Id: Id37d88bad93d154826e132dc243e24daf27d3905

Former-commit-id: 9ef0692a504a73b989208bf2a49ddcb82fbde549