Look for the installer in all users if DELETE_ALL_USERS is requested
If a caller  passes the DELETE_ALL_USERS flag, search for the installer
app in every user, since the uninstall request may initiate from a user
where the package is not installed.

For example, if an app is installed only on user10, and a caller on
user0 is initiating the uninstall (to uninstall from all users),
looking for for the installer package* only on the initiating user
(i.e user 0) will lead to an exception. Thus, loop through all users to
find the installer

(*) We search for the installer to determine if the caller pkg is also
the installer pkg, to allow silent uninstall

Bug: 384544054
Test: TBD
Flag: EXEMPT. Bug fix only
Change-Id: Ibc893fbf32bcef0601cf57cb798f582635ab00c7

Former-commit-id: ff21533e7386987fe668e9d4cee20dc2d6b99add