Do not unparcel LazyValue in collectExtraIntentKeys method
collectExtraIntentKeys should only care about intents created in
this process, which would not be parceled when put in the extra
bundle. Unpparcel a LazyValue may cause issues when an app later
sets class loader to something else and prefer to unparcel it with
this new class loader.

Bug: 382633789
Test: manual test
Flag: EXEMPT bug fix
Change-Id: I4e14b255e8356a94816c38e8d0e9fa69ac38e6fb

Former-commit-id: 7b342c8a83825e6b8d35613159133ed9e0597480