[1/n] Freeze rotation for SCM activities in desktop mode
When a SCM activity is in desktop mode, if the device is rotated don't
inherit new parent rotation. This essentially freezes the rotation of
the activity to that when it first enters SCM in desktop mode.

Replace `appCompatDisplayInsets.mIsFloating` with
`newParentConfiguration.windowConfiguration.tasksAreFloating()` to
respect current window mode and enfoce this behaviour.

Flag: NONE (bug fix)
Test: manual testing
Fixes: 368104917
Change-Id: I3964e1df30ce4a1d5b04a5159d89fee938acdb9b

Former-commit-id: ce5c8fcde2e0d94f6e0bc2c6deaeedd1e63956d3