Group session MediaItems together in OutputSwitcher
If multiple routes are selected, they will be initially grouped
together, with a carat that allows expansion. The volume seekbar and
icon controls the sesion volume, and the name of the entry is the
session name.

There must already be an existing session before opening the output
switcher for them to be grouped. If a device is added to the session, it
will maintain its existing position until the output switcher is
reopened.

Bug: 388347018
Test: Tested locally, atest
Flag: com.android.media.flags.enable_output_switcher_session_grouping
Change-Id: I7763783ddf4cf66d35dbe34a5f7620fa82ead7cc

Former-commit-id: 76a75f865eadb26afff8432dcfd6d15a838b41c2