Fix SHOW_WHEN_LOCKED splitscreen bug
In ag/29956640, the lock/wake behavior for splitscreen was changed to fix a camera bug. At the time, I think the SHOW_WHEN_LOCKED behavior was working as well. However, the logic is no longer working in the same way today, and the SHOW_WHEN_LOCKED issue has regressed. It seems to be because we are receiving TRANSIT_TO_FRONT changes now, instead of TRANSIT_TO_BACK as previously expected.

To fix this, I changed the logic to account for both TRANSIT_TO_BACK and TRANSIT_TO_FRONT changes -- both apps must receive a TO_BACK or TO_FRONT in the same transition, or else we will break split. This fixes the regression.

Fixes: 387227528
Test: Repro steps do not repro anymore
Flag: EXEMPT bugfix
Change-Id: Ib833c78032545032797593578ac61478b695d558

Former-commit-id: ea5fb8087eece78df3f6caf79982c32971784b60