Add new API to check if a custom gesture trigger is available
In order to notify user that a key combination is already taken,
we need another API to check if a input gesture trigger is
available. This is a new UX request allowing user to pre-emptively
know if adding a custom gesture is allowed for a trigger or not.

Bug: 381063978
Test: atest InputTests
Flag: EXEMPT internal API addition only
Change-Id: Id579fa179acabc783ab6ca61bf34b8b74dc5e1af

Former-commit-id: af01f6dff2792ce8c4b1817ce05964c96b1477a2