Force unparcel values when collecting intent keys on server side.
Unparceling extra values in an intent on the client side has some
undesired side effect. So we fixed it in ag/31181657. But this would
break collecting intent keys on the server side since everything
is parceled on the server side. Fix it by passing a boolean flag.

Bug: 382633789
Test: manual test
Flag: EXEMPT bug fix
Change-Id: I57590b67d789b43ed08bac53de52efc3031b26a6

Former-commit-id: bf614e3f2b6af65e6380127de097b52ff51ab8bc