Do not allow non-system apps to provide unverified attributions
Some apps (the shell, system server, etc) are exempt from the
requirement that attribution tags be registered. However, in the proxy
case, the tag provied by the proxy app is trusted if the proxied app is
one of these exemptions. We should only trust these tags if the proxy
app is a system app.

This CL also adds a second restriction check when a restriction is
removed, to verify that an op is free of all restrictions, before
resuming a started op

Bug: 375623125
Test: upcoming
Flag: EXEMPT: See bug
Merged-In: I6a7b0a24359097c0ea2f52cc69637d929a931b4f
Change-Id: I6a7b0a24359097c0ea2f52cc69637d929a931b4f

Former-commit-id: b645679d1d40d05944527eee4b5fe5cae0c3fc18