Always enable sparse encoding if minSdk >= 32
Sparse encoding has been enabled in the framework since T QPR2 with no major issues. It should be safe to enable by default for any app with minSdk >= 32 (S_V2) which is when the last fix for sparse encoding was made.

This CL makes the `--enable-sparse-encoding` flag a no-op. As such, we are also bumping the version to 2.20.

Bug: 302179337
Change-Id: Ia886a0da9d61175d00c24e30f4e263f7e5df511a
Test: build + boot on Pixel 6 Pro
Flag: EXEMPT command line tool

Former-commit-id: c9c3150896e1b637df6c8ecc485cb0e3384448c7