Do not block intetn in a top level ClipData from being launched
If an app that supports drag-and-drop and its drop target performs
launching the intent passed through the ClipData, the intent is
blocked due to missing creator token error. The root cause is the
ClipData.Item.getIntent() method calls maybeMarkAsMissingCreatorToken
even if the ClipData is a top level object passed to the application.
In such case, the ClipData won't go through the collect keys and
add creator token to the intent. But it will be marked as missing
creator token when being launched. To fix it, follow the example of
how we handle intent's extra bundle. Only clipData inside an intent
would be marked as need token verification. And if it is not marked,
do not call maybeMarkAsMissingCreatorToken in Item.getIntent()

Bug: 388879274
Test: ActivityManagerServiceTest#testCanLaunchClipDataIntent
Flag: EXEMPT bug fix
Change-Id: I88a904c96e0cf646dd25070bc12fbc344967feb3

Former-commit-id: 965db179f4310d01b1d3fed52610f4e3a6df869b