If android version is different & app has SAW, auto-grant bubbles
We've had developer feedback around this not applying to all versions
of android.

To address this feedback, this CL adds an attribute to save which
android version was last checked for the bubbles auto-grant. If
the current android version doesn't match whats saved, perform the
check for SAW again.

Additionally checks if the user has locked the field / changed the
bubble preference before, if so, don't modify it.

Adds a test that verifies we don't call app-ops if the version is the
same and adds a test for the versions being different.

Existing tests verify that the pref isn't overridden if it's locked.

Flag: EXEMPT bug fix
Test: atest PreferencesHelperTest
Bug: 372294507
Change-Id: Ic6c879ee4401b122c2bf9680be9dfe22d6baf15f



Former-commit-id: 2e6b585cf2f01be48d9f1b9286fb45db5e91b755