Remove the version check for opting apps with SAW into bubbles
We've had developer feedback around this not applying to all versions
of android, so modifying the version check to be >= instead of =.

Additionally checks if the user has locked the field / changed the
bubble preference before, if so, don't modify it.

Added a test that has the current version in XML and verifies the
bubble preferences is updated.

Existing tests verify that the pref isn't overridden if it's locked.

Flag: EXEMPT bug fix
Test: atest PreferencesHelperTest
Bug: 372294507
Change-Id: I234e633d651066ac9ff5b8c552db7719ff176512



Former-commit-id: b4fd0a4ad60caa70ec759c71f09ee483e9c6d5e0